@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: '>=2.19.4 <4.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -21,6 +21,9 @@ dependencies:
 
   english_words: ^4.0.0
   provider: ^6.0.0
+  flutter_svg: ^2.2.1
+  xml: ^6.5.0
+  vector_math: ^2.1.2
 
 dev_dependencies:
   flutter_test:
