@@ -1,41 +1,35 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'shopping_list_model.dart';
+part of 'layout_model.dart';
 
 // **************************************************************************
 // TypeAdapterGenerator
 // **************************************************************************
 
-class ShoppingListModelAdapter extends TypeAdapter<ShoppingListModel> {
+class LayoutModelAdapter extends TypeAdapter<LayoutModel> {
   @override
-  final int typeId = 0;
+  final int typeId = 2;
 
   @override
-  ShoppingListModel read(BinaryReader reader) {
+  LayoutModel read(BinaryReader reader) {
     final numOfFields = reader.readByte();
     final fields = <int, dynamic>{
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
-    return ShoppingListModel(
-      id: fields[0] as String,
-      name: fields[1] as String,
-      productIds: (fields[2] as List).cast<String>(),
-      showInOtherView: (fields[3] as bool?) ?? false,
+    return LayoutModel(
+      layoutHash: fields[0] as String,
+      layoutSvg: fields[1] as String,
     );
   }
 
   @override
-  void write(BinaryWriter writer, ShoppingListModel obj) {
+  void write(BinaryWriter writer, LayoutModel obj) {
     writer
-      ..writeByte(4)
-      ..writeByte(0)
-      ..write(obj.id)
-      ..writeByte(1)
-      ..write(obj.name)
       ..writeByte(2)
-      ..write(obj.productIds)
-      ..writeByte(3)
-      ..write(obj.showInOtherView);
+      ..writeByte(0)
+      ..write(obj.layoutHash)
+      ..writeByte(1)
+      ..write(obj.layoutSvg);
   }
 
   @override
@@ -44,7 +38,7 @@ class ShoppingListModelAdapter extends TypeAdapter<ShoppingListModel> {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is ShoppingListModelAdapter &&
+      other is LayoutModelAdapter &&
           runtimeType == other.runtimeType &&
           typeId == other.typeId;
 }
